restructuring and handling odoo sale.order
